# contracts/openapi/equipment.json
{
  "components" : {
    "schemas" : {
...
      "ClientEquipmentRegisterRequest" : {
        "properties" : {
          "codigoInterno" : {
            "description" : "El codigo que el cliente le pone a su maquina",
            "maxLength" : 50,
            "minLength" : 0,
            "type" : "string"
          },
          "fechaCompra" : {
            "format" : "date",
            "type" : "string"
          },
          "idModelo" : {
            "format" : "uuid",
            "type" : "string"
          },
          "numeroInventario" : {
            "maxLength" : 50,
            "minLength" : 0,
            "type" : "string"
          },
          "proveedor" : {
            "description" : "Quien vendio esta maquina",
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          },
          "serie" : {
            "maxLength" : 50,
            "minLength" : 0,
            "type" : "string"
          },
          "valorCompra" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "idModelo", "serie" ],
        "type" : "object"
      },
      "ClientEquipmentResponse" : {
        "properties" : {
          "codigoInterno" : {
            "type" : "string"
          },
          "estadoActivo" : {
            "type" : "boolean"
          },
          "fechaCompra" : {
            "format" : "date",
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "idAreaServicio" : {
            "format" : "uuid",
            "type" : "string"
          },
          "idModelo" : {
            "format" : "uuid",
            "type" : "string"
          },
          "numeroInventario" : {
            "type" : "string"
          },
          "proveedor" : {
            "type" : "string"
          },
          "serie" : {
            "type" : "string"
          },
          "valorCompra" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "ClientEquipmentUpdateRequest" : {
        "properties" : {
          "codigoInterno" : {
            "description" : "El codigo que el cliente le pone a su maquina",
            "maxLength" : 50,
            "minLength" : 0,
            "type" : "string"
          },
          "fechaCompra" : {
            "format" : "date",
            "type" : "string"
          },
          "numeroInventario" : {
            "maxLength" : 50,
            "minLength" : 0,
            "type" : "string"
          },
          "proveedor" : {
            "description" : "Quien vendio esta maquina",
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          },
          "valorCompra" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "EquipmentCreateRequest" : {
        "properties" : {
          "idMarca" : {
            "format" : "uuid",
            "type" : "string"
          },
          "idTipoEquipo" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "required" : [ "idMarca", "idTipoEquipo" ],
        "type" : "object"
      },
      "EquipmentErrorResponse" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "details" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "message" : {
            "type" : "string"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "EquipmentResponse" : {
        "properties" : {
          "estadoActivo" : {
            "type" : "boolean"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "idMarca" : {
            "format" : "uuid",
            "type" : "string"
          },
          "idTipoEquipo" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "GlobalErrorResponse" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "details" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "message" : {
            "type" : "string"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "bearerFormat" : "JWT",
        "description" : "Alternativa: pegar un token obtenido por otro medio",
        "scheme" : "bearer",
        "type" : "http"
      },
      "keycloak-oauth" : {
        "description" : "Inicia sesion en Keycloak y usa el token en cada peticion",
        "flows" : {
          "authorizationCode" : {
            "authorizationUrl" : "http://localhost:8080/realms/malphasos-realm/protocol/openid-connect/auth",
            "scopes" : {
              "email" : "Correo electronico",
              "openid" : "Identificacion del usuario",
              "profile" : "Datos basicos del perfil"
            },
            "tokenUrl" : "http://localhost:8080/realms/malphasos-realm/protocol/openid-connect/token"
          }
        },
        "type" : "oauth2"
      }
    }
  },
  "info" : {
    "description" : "API REST para la gestión de mantenimientos preventivos y clientes",
    "license" : {
      "name" : "GNU GPL v3",
      "url" : "https://www.gnu.org/licenses/gpl-3.0.html"
    },
    "title" : "MalphasOS API",
    "version" : "v0.0.1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/v1/api/client-equipments" : {
      "get" : {
        "operationId" : "getAll_6",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ClientEquipmentResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EquipmentErrorResponse"
                }
              }
            },
            "description" : "Bad Request"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EquipmentErrorResponse"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalErrorResponse"
                }
              }
            },
            "description" : "Conflict"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalErrorResponse"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Listar todas las unidades",
        "tags" : [ "ClientEquipment" ]
      }
    },
    "/v1/api/client-equipments/{id}" : {
      "delete" : {
        "description" : "No la borra: la deja inactiva.",
        "operationId" : "decommission",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EquipmentErrorResponse"
                }
              }
            },
            "description" : "Bad Request"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EquipmentErrorResponse"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalErrorResponse"
                }
              }
            },
            "description" : "Conflict"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalErrorResponse"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Dar de baja una unidad",
        "tags" : [ "ClientEquipment" ]
      },
      "get" : {
        "operationId" : "getById_3",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ClientEquipmentResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EquipmentErrorResponse"
                }
              }
            },
            "description" : "Bad Request"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EquipmentErrorResponse"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalErrorResponse"
                }
              }
            },
            "description" : "Conflict"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalErrorResponse"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Obtener una unidad por su identificador",
        "tags" : [ "ClientEquipment" ]
      },
      "patch" : {
        "description" : "Los campos ausentes conservan su valor.",
        "operationId" : "update_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ClientEquipmentUpdateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ClientEquipmentResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EquipmentErrorResponse"
                }
              }
            },
            "description" : "Bad Request"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EquipmentErrorResponse"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalErrorResponse"
                }
              }
            },
            "description" : "Conflict"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalErrorResponse"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Corregir los datos de compra de una unidad",
        "tags" : [ "ClientEquipment" ]
      }
    },
...
    "/v1/api/client-equipments/{id}/service-area/{idAreaServicio}" : {
      "patch" : {
        "description" : "El area de destino debe estar activa y pertenecer al mismo cliente que la unidad. Entre sedes de ese cliente si se puede; cruzar de cliente responde 409.",
        "operationId" : "relocate",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "idAreaServicio",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ClientEquipmentResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EquipmentErrorResponse"
                }
              }
            },
            "description" : "Bad Request"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EquipmentErrorResponse"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalErrorResponse"
                }
              }
            },
            "description" : "Conflict"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalErrorResponse"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Trasladar una unidad a otra area de servicio",
        "tags" : [ "ClientEquipment" ]
      }
    },
...
    "/v1/api/equipments" : {
      "get" : {
        "description" : "Con el parametro idMarca, solo las de esa marca.",
        "operationId" : "getAll_2",
        "parameters" : [ {
          "description" : "Filtra por marca",
          "in" : "query",
          "name" : "idMarca",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/EquipmentResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EquipmentErrorResponse"
                }
              }
            },
            "description" : "Bad Request"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EquipmentErrorResponse"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalErrorResponse"
                }
              }
            },
            "description" : "Conflict"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalErrorResponse"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Listar asociaciones",
        "tags" : [ "Equipment" ]
      },
      "post" : {
        "operationId" : "create_2",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EquipmentCreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EquipmentResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EquipmentErrorResponse"
                }
              }
            },
            "description" : "Bad Request"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EquipmentErrorResponse"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalErrorResponse"
                }
              }
            },
            "description" : "Conflict"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalErrorResponse"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Registrar que una marca fabrica un tipo de equipo",
        "tags" : [ "Equipment" ]
      }
    },
    "/v1/api/equipments/{id}" : {
      "delete" : {
        "description" : "No la borra: la deja inactiva.",
        "operationId" : "deactivate_4",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EquipmentErrorResponse"
                }
              }
            },
            "description" : "Bad Request"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EquipmentErrorResponse"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalErrorResponse"
                }
              }
            },
            "description" : "Conflict"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalErrorResponse"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Retirar una asociacion",
        "tags" : [ "Equipment" ]
      },
      "get" : {
        "operationId" : "getById_5",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EquipmentResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EquipmentErrorResponse"
                }
              }
            },
            "description" : "Bad Request"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EquipmentErrorResponse"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalErrorResponse"
                }
              }
            },
            "description" : "Conflict"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalErrorResponse"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Obtener una asociacion por su identificador",
        "tags" : [ "Equipment" ]
      }
    },
...
    "/v1/api/service-areas/{idAreaServicio}/equipments" : {
      "get" : {
        "operationId" : "getByServiceArea",
        "parameters" : [ {
          "description" : "Identificador del area",
          "in" : "path",
          "name" : "idAreaServicio",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ClientEquipmentResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EquipmentErrorResponse"
                }
              }
            },
            "description" : "Bad Request"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EquipmentErrorResponse"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalErrorResponse"
                }
              }
            },
            "description" : "Conflict"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalErrorResponse"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Inventario de un area de servicio",
        "tags" : [ "ClientEquipment" ]
      },
      "post" : {
        "description" : "El area debe estar activa y el modelo no puede estar retirado.",
        "operationId" : "register",
        "parameters" : [ {
          "in" : "path",
          "name" : "idAreaServicio",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ClientEquipmentRegisterRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ClientEquipmentResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EquipmentErrorResponse"
                }
              }
            },
            "description" : "Bad Request"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EquipmentErrorResponse"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalErrorResponse"
                }
              }
            },
            "description" : "Conflict"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalErrorResponse"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Incorporar una unidad al inventario de un area",
        "tags" : [ "ClientEquipment" ]
      }
    }
  },
  "security" : [ {
    "keycloak-oauth" : [ ]
  }, {
    "bearer-jwt" : [ ]
  } ],
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ],
...
}